Error fix for C file "5-rev_string.c"

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -14,8 +14,8 @@
 
 void rev_string(char *s)
 {
-	int i;
-	int t;
+	size_t int i;
+	size_t int t;
 	size_t len;
 	len = strlen (s);
